requirements.txt
astroid==2.1.0
autopep8==1.4.3
Django==2.1.3
django-filter==2.0.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.4.0
PyJWT==1.7.1
pylint==2.2.2
pytz==2018.7
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
